Update ReadResult.Data to return typed rows
